Change floats to doubles in capacity and restocking calculations

+++ Assets/Src/FirstLight/Game/Configs/AdjectiveDataConfigs.cs
@@ -9,7 +9,7 @@ namespace FirstLight.Game.Configs
 	public struct AdjectiveDataConfig
 	{
 		public EquipmentAdjective Adjective;
-		public float PoolCapacityModifier;
+		public double PoolCapacityModifier;
 	}
 
 	/// <summary>
